Refactoring CertificateDTO, add new CertificateDTO, fix bugs

+++ src/main/java/com/epam/esm/dao/CertificateDao.java
@@ -1,14 +1,13 @@
 package com.epam.esm.dao;
 
 import com.epam.esm.entity.GiftCertificate;
-import com.epam.esm.entity.GiftCertificateDataTransferObject;
 
 import java.util.List;
 
 public interface CertificateDao {
     public List<GiftCertificate> findAllCertificates();
     public GiftCertificate findCertificateById(int id);
-    public int createNewCertificate(GiftCertificate certificate);
+    public long createNewCertificate(GiftCertificate certificate);
     public void updateCertificate(GiftCertificate certificate, int id);
     public void deleteCertificate(int id);
     public  int findcertificateIdByCertificateInformation(GiftCertificate certificate);
