+++ src/main/java/com/epam/esm/dao/CertificateTagDao.java
@@ -2,5 +2,5 @@ package com.epam.esm.dao;
 
 public interface CertificateTagDao {
 
-    public void createNewCertificateTagRelation(int certificateId, int tagId);
+    public void createNewCertificateTagRelation(long certificateId, long tagId);
 }
